requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
idna==3.3
iniconfig==1.1.1
loguru==0.5.3
more-itertools==8.11.0
packaging==21.2
pluggy==0.13.1
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
selenium==3.14.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
wcwidth==0.2.5
webdriver-manager==3.5.2
win32-setctime==1.0.3